Updated README.md
Wind speed output as rounded whole numbers rather than with decimal values as in the original design. Fixes missing wind speed values to accurately reflect that shown on console. For example, no longer shows 2.2 mph for both 1 mph and 3 mph. 

Average wind speed is for 10 minutes as shown on console rather than 60 minutes. Fixed. 
Changed 1 hour rain total in Current Conditions to 3 hours so as not to duplicate that already shown in the rain module.